+++ blueprints/cis/level2/locals.tf
@@ -4,4 +4,3 @@ locals {
   cis_level      = "level2"
   name_prefix    = "${var.org}-${var.environment}-${var.region_key}"
 }
-